update the btree even when the file was found in the cache
when we scan the files we first put a Entry structure in the index and
only later we consult the cache to see if we've already chunkified it.

The issues is that we have to update it even in the case the entry was
found in the cache, otherwise we leave the Objects field to nil in the
backup and the file content are not available.