fix some ls weirdnesses
Make sure the path is cleaned or the WalkDir callback breaks.

Do not print the named directory in its output.  In ls(1) you can
achieve it only by passing -d, something which we don't have (and
doesn't seem terribly useful in this context.)

Then, unless -recursive is passed, print only the entry name, not
not its full path.