vfs: add another btree for extended attributes
Since these can be varying in size (on darwin for example), we're
tracking their content in an object like it were a file itself but to
avoid cluttering the "real" VFS, we're keeping these entries in their
own tree which the VFS package take care of.

The interface to handle xattrs will be improved in subsequent commits.

w/ gilles@