state: A few changes to how we key off DeltaEntry.
* To be able to allow concurrent backups (possibily leading to same Blob
in different packfiles) we need to add the Packfile id at the end of
each blob Key, thereby allowing multiple Blob keys.

* This turns Retrieval/Existance check into something a bit harder, and
also slower but there is no other way to do this properly.

* This gives us the ability to also not even consider any blob entry
which packfile is not in our know state packfiles list, allowing
atomical publication of entire packfiles' worth of blob.