summaries: Fix accounting of directories.
* We use to increment the directory count in UpdateWithFileSummary()
(the name should be a hint that this is wrong :D). Now that we iterate
in two steps (files then directories) for each directories, we need to
account for that in the directories branch.

* Note: This will also count the root parent dirs (the resolution up to
the root that is not part of the path being backed up), I don't know if
this desirable.

* Note2: Maybe tehre are more stats we want to account for and we should
have a UpdateWithDirSummary()?