api: serve all text/* as text/plain
niluje@ spotted that text/xml was still interpreted; so let's use the
big hammer and serve all text/* stuff as text/plain to stay on the safe
side.

The intent is to always show the content of the files in the UI, so we
have to prevent the browser from trying to be 'smart'.