cmd: Remove the RPC interface.
* The RPC interface relied on the Name() method to select which commands
could get sent over the wire. Retire the Name method and the whole
interface so that we can use Subcommand everywhere including over the
wire.

* In place use the existing Flags mechanism to tell the agent which
command can be used over the wire or not.

* This needs a bit of tweaking to the Lookup method, as we rely on it to
set the flags on the object, meaning we don't return a factory anymore
but a created (and updated) object. This also means no one should use
the real objects without a Lookup() before.