attempt to please CodeQL in the less horrible way possible
CodeQL is "smart" enough to find out that the MAC passed to, for e.g.,
buckets.Put might come from the http server, and so it's user
controlled.  Great.  Except it completely misses the point that the
MAC gets serialized in hex format and so it's only made up of [A-Z].
The constructed path can't escape buckets.path in any way.