storage/s3: Pool buffers for PutPackfile.
* We were reported plakar getting OOM killed on somewhat medium sized
backups done on a S3 repository. Turns out since we don't know the size
of the file we PutObject, the library uses a default partition size of
512MB. Since we do up to at most 2*CPUs+1 PutPackfile at once, on
machines with a lot of cores and relatively small memory we overcommit
quickly.

* minio doesn't let you provide a "imprecise" size to better size the
default buffer, so we do need to compute the size ourselves. In order to
do that we go through a ByteBuffer and do the copy ourselves. It's
always faster than them doing the multipart dance, and we can pool it
too.