importer/fs: add knob to not traverse filesystems
This adds a knob for the fs importer to not traverse filesystems
during a backup.  The idea is that we keep track of the device number
for the root of the backup and we skip all the directories that belong
to a different device.

The new knob is called dont_traverse_fs and at the moment has to be
specified by creating a "remote" config stanza with a trueish value
("1", "t", "T", "true", "TRUE" or "True").