split NormalizePath per-OS
Default settings on macos and Windows filesystem are to have every
directory entry being case-insensitive, i.e. FOO.txt and foo.txt
on windows are the same file.  On unices (whether darwin is a real
UNIX is something I personally disagree) the filesystem is
case-sensitive (heck, the filesystem doesn't even have the concept
of 'casing' here.)

So, move the existing NormalizePath only on windows and darwin,
since on unix we might have both 'foo' and 'FOO' as path component
and they are different.

Technically speaking on windows is possible to set per-directory
case-sensitiveness, but for now ignore that detail.